Add a ClassificationResults class
Creates the inferencedata module for classes that manage
inference data. Adds a subclass of arviz.InferenceData for
managing the results of classification inference runs.

The class can be instantiated from normal InferenceData objects
and from json files. It determines which reactor class is the
result of the inference from the posterior of the class variable.

Adds tests for the ClassificationResults class.