Add AccuracyMeasure classes to metrics
Subclasses of AccuracyMeasure are intended to compare a set of predicted
values with a corresponding set of true values to assess the accuracy of
a prediction framework.

Three measures are already implemented: R-squared, mean absolute
percentage error and the root mean squared error.

Tests for these classes are also added.